Add a script for building a .deb binary package
Add a new script (package/debian.pl) which compiles and strips the
application and builds it into an installable .deb package for Debian
based Linux distros.

This calls another new script (package/manpage.pl) which converts the
markdown help page into a man page for inclusion in Linux packages.
Also updated the help page so that it is no-longer Windows specific and
fits better with other man pages on Linux, but without it being less
useful elsewhere.

The package also adds a launcher entry and associates .bbg files with
beebview.