Move author copyright to individual files
Instead of trying to handle the list of authors separately to the file
copyright, remove AUTHORS.md and change the copyright in all the files
to the contributor names.  This helps to keep the author copyright with
the file content if any of the files are used elsewhere.

This required making some minor adjustments where the build collated a
list of author names and copyright dates from AUTHORS.md to show in the
About dialog, Windows .exe properties and .deb copyright - instead just
show the dates with a hardcoded 'from' year and the year of the latest
commit as the 'to'.

While updating the copyright headers, also correct a couple of the dates
as these were incorrect and take the opportunity to reduce the size of
the file headers by removing the GPL boilerplate from each file and
replacing it with a SPDX-License-Identifier header line.