Enable and add a range of Overcommit hooks
Enable some out of the box Overcommit rules for checking trailing
whitespace and running yamllint and shellcheck.  Add four custom hooks
which ensure files end with a single newline, are encoded as UTF-8
without a BoM and have up-to-date copyright notices and that perl
scripts pass checking with perlcritic.

Also fix the shellcheck and perlcritic violations found by those checks
and add a .perlcriticrc file to specify which checks to run.  At the
same time, rename the linux packaging and manpage generation scripts for
clarity.