Enable running build & package in the CI workflow
Remove the package_* scripts and hook up the CI workflow to perform a
build and package (in addition to the existing Overcommit run).

Installing the dependencies is straightforward under Ubuntu but as the
Actions Windows runner images don't have Qt pre-installed, use
jurplel/install-qt-action to fetch, install and configure the latest Qt5
release.