Use Qt command to install runtime deps for Windows
Instead of gathering the runtime dependencies for Windows ourselves,
switch to using qt_generate_deploy_app_script instead.  This was only
introduced in Qt version 6.3 so couldn't be used previously.

Set the NO_TRANSLATIONS option to suppress an error when the install
script is called and set NO_COMPILER_RUNTIME as Qt doesn't currently
seem to handle this correctly with MSVC.

This also required bumping the cmake_minimum_required version to 3.14 to
enable the generator expression used in the install SCRIPT argument.